+++ MercadoPagoSDK/MercadoPagoSDK/Services/Models/PXRemoteAction.swift
@@ -7,10 +7,16 @@
 
 import Foundation
 
-struct PXRemoteAction: Codable {
+@objcMembers
+public class PXRemoteAction: NSObject, Codable {
 
     let label: String
     let target: String?
+    
+    public init(label:String, target:String?) {
+        self.label = label
+        self.target = target
+    }
 
     enum CodingKeys: String, CodingKey {
         case label